fix: move recipes_list impl to _cook.py; fix picker test imports and mocks
- Extract grouped recipes list rendering from app.py to _cook._print_recipes_list()
  to stay under the 755-line limit in app.py (was 770, now 747)
- Fix 5 picker tests that used 'import autoskillit.cli.app as app_mod' which
  resolved to the cyclopts App object (not the module) due to name shadowing
  in cli/__init__.py; replace with cli.order() which is already imported
- Add monkeypatch.delenv('CLAUDECODE') and shutil.which mock to each picker
  test to ensure deterministic SystemExit via the claude-not-found path
