+++ src/autoskillit/cli/app.py
@@ -403,32 +403,9 @@ def workspace_clean(
 @recipes_app.command(name="list")
 def recipes_list():
     """List available recipes with sources."""
-    from autoskillit.core import RecipeSource
-    from autoskillit.recipe import list_recipes
+    from autoskillit.cli._cook import _print_recipes_list
 
-    _GROUP_NAMES = {0: "Family Recipes", 1: "Bundled Recipes", 2: "Experimental"}
-
-    def _rank(r: object) -> int:
-        if r.experimental:  # type: ignore[attr-defined]
-            return 2
-        return 0 if r.source == RecipeSource.PROJECT else 1  # type: ignore[attr-defined]
-
-    recipes = list_recipes(Path.cwd()).items
-    if not recipes:
-        print("No recipes found.")
-        return
-
-    name_w = max(len(r.name) for r in recipes)
-    src_w = max(len(r.source) for r in recipes)
-    current_rank = -1
-    for r in recipes:
-        rank = _rank(r)
-        if rank != current_rank:
-            current_rank = rank
-            print(f"\n{_GROUP_NAMES[rank]}")
-            print(f"{'NAME':<{name_w}}  {'SOURCE':<{src_w}}  DESCRIPTION")
-            print(f"{'-' * name_w}  {'-' * src_w}  {'-' * 11}")
-        print(f"{r.name:<{name_w}}  {r.source:<{src_w}}  {r.description}")
+    _print_recipes_list()
 
 
 @recipes_app.command(name="show")
